chore: split GraphQL functions into separate files, clean up interfaces / types
