+++ packages/ingestor/src/additional-objects/dynamicObjects.ts
@@ -2,7 +2,7 @@
  * Specs for dynamic objects that are created by the ingestor outside of Airtable
  * (Usually you'd create these through the CMS)
  */
-import { DynamicObjectConfig } from "../interfaces";
+import { DynamicObjectConfig } from "../lib/interfaces";
 
 export const quentinTarantinoMovies: DynamicObjectConfig = {
   name: "quentin-tarantino-movies",
