+++ packages/ingestor/src/lib/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from "./airtable";
+export * from "./config";
+export * from "./classic";
+export * from "./saas";
