+++ packages/ingestor/src/lib/skylark/classic/create.test.ts
@@ -13,7 +13,7 @@ import {
   ApiEntertainmentObjectWithAirtableId,
   DynamicObjectConfig,
   Metadata,
-} from "../../../interfaces";
+} from "../../interfaces";
 import {
   convertAirtableFieldsToSkylarkObject,
   createOrUpdateAirtableObjectsInSkylark,
