+++ packages/ingestor/src/lib/skylark/classic/utils.test.ts
@@ -1,4 +1,4 @@
-import { Metadata } from "../../../interfaces";
+import { Metadata } from "../../interfaces";
 import {
   getScheduleUrlsFromMetadata,
   removeUndefinedPropertiesFromObject,
