+++ packages/ingestor/src/lib/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./classic";
+export * from "./saas";
